### 2023-02-27 ReSurfEMG/ReSurfEMG#223 @drcandacemakedamoore
fix DOIs on paper
MISSING DOIs

- 10.1007/978-3-642-34546-3_71 may be a valid DOI for title: Towards Improving the Usability of Electromyographic Interfaces

INVALID DOIs

- https://doi.org/10.1016/0013-4694(86)90163-X is INVALID because of 'https://doi.org/' prefix
- https://doi.org/10.1046/j.1365-2842.1998.00242.x is INVALID because of 'https://doi.org/' prefix
