### 2023-02-27 ReSurfEMG/ReSurfEMG#223 @drcandacemakedamoore
@ElineOppersma  can you fix these since you have a pending PR on the paper anyways?
