+++ paper/paper.bib
@@ -6,7 +6,7 @@ number = {6},
 pages = {511-519},
 year = {1996},
 issn = {0924-980X},
-doi = {https://doi.org/10.1016/S0921-884X(96)95190-5},
+doi = {10.1016/S0921-884X(96)95190-5},
 url = {https://www.sciencedirect.com/science/article/pii/S0921884X96951905},
 author = {Paul W. Hodges and Bang H. Bui},
 keywords = {Electromyogram (EMG), Onset, Timing, Processing, Analysis},
@@ -31,7 +31,7 @@ number = {4},
 pages = {378-380},
 year = {1986},
 issn = {0013-4694},
-doi = {https://doi.org/10.1016/0013-4694(86)90163-X},
+doi = {10.1016/0013-4694(86)90163-X},
 url = {https://www.sciencedirect.com/science/article/pii/001346948690163X},
 author = {M Lidierth},
 keywords = {electromyography, locomotion, automatic data processing, computer},
@@ -47,7 +47,7 @@ journal = {Journal of Oral Rehabilitation},
 volume = {25},
 number = {5},
 pages = {365-369},
-doi = {https://doi.org/10.1046/j.1365-2842.1998.00242.x},
+doi = {10.1046/j.1365-2842.1998.00242.x},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1046/j.1365-2842.1998.00242.x},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1046/j.1365-2842.1998.00242.x},
 abstract = {SUMMARY A method of automated detection of onset and termination of rhythmic muscle activity in electromyograms (EMGs) is presented. A threshold level in the EMG is computed, such that amplitudes in the EMG signal exceeding this level indicate muscle activity. The threshold level is determined using a statistical criterion based on the amplitude distribution of the entire EMG signal. The working of the method is illustrated with EMG signals recorded from chewing muscles. EMG signals with a good as well as a worse signal-to-noise ratio are presented. The method can be used for any EMG signal containing cyclic bursts of activity and thus may be applied in studies on rhythmic movements, such as chewing, walking and breathing. An automated method of EMG burst detection has the advantage that large amounts of EMG data can be easily and objectively processed.},
