### 2023-02-28 #5192 @AJQuinn
I was about to ask about the DOIs but looks like you're already on it @drcandacemakedamoore! - https://github.com/ReSurfEMG/ReSurfEMG/issues/223

Let me know once you have a fix.
