### 2023-03-04 281a8b3
add variability function
