### 2023-03-04 5d7e4dc
add testing for variability maker
