### 2023-03-06 #5192 @AJQuinn
@drcandacemakedamoore thanks for the fix. I'll let you know if the remaining DOI query needs further checking.
