+++ paper/paper.bib
@@ -143,3 +143,11 @@ journal = {Journal of Open Source Software} }
   author={Milivojevi{\'c}, Milan S and Gavrovska, Ana and Reljin, Irini S and others},
   year={2017}
 }
+
+@Misc{Carreiras2015biosppy,
+  author = {Carlos Carreiras and Ana Priscila Alves and Andr\'{e} Louren\c{c}o and Filipe Canento and Hugo Silva and Ana Fred and others},
+  title = {{BioSPPy}: Biosignal Processing in {Python}},
+  year = {2015--},
+  url = "https://github.com/PIA-Group/BioSPPy/",
+  note = {[Online; accessed <today>]}
+}
