+++ paper/paper.md
@@ -43,7 +43,7 @@ ReSurfEMG code allows for analysis of various parameters in respiratory surface 
 
 # State of the field
 
-Respiratory surface EMG signals can be processed, with a limited group of algorithms, by several existing libraries which deal with EMG signals in general. However, none of these libraries have been extended publicly to cover all the specifics of the respiratory signal. BioSPPy [@milivojevic2017python], and NeuroKit [@Makowski2021] both cover processing for various electrophysiological signals, and have modules for general EMG processing. Unfortunately, these modules do not have code specific for respiratory surface EMG. pyemgpipline [@Wu2022] is specific to EMG, but not to respiratory EMG. Without functions specific to respiratory EMG, researchers must code themselves the functions for extracting even basic parameters reported in current literature, such as area under curve. Aside from lacking transparency in clinical scientific literature, this is beyond the technical skills of many researchers in the field, and hinders the quality and the reproducibility of the results. Therefore, ReSurfEMG bridges the gaps in clinical researcher abilities, allowing researchers at a low level of technical skill to analyze respiratory EMG.  
+Respiratory surface EMG signals can be processed, with a limited group of algorithms, by several existing libraries which deal with EMG signals in general. However, none of these libraries have been extended publicly to cover all the specifics of the respiratory signal. BioSPPy [@Carreiras2015biosppy], and NeuroKit [@Makowski2021] both cover processing for various electrophysiological signals, and have modules for general EMG processing. Unfortunately, these modules do not have code specific for respiratory surface EMG. pyemgpipline [@Wu2022] is specific to EMG, but not to respiratory EMG. Without functions specific to respiratory EMG, researchers must code themselves the functions for extracting even basic parameters reported in current literature, such as area under curve. Aside from lacking transparency in clinical scientific literature, this is beyond the technical skills of many researchers in the field, and hinders the quality and the reproducibility of the results. Therefore, ReSurfEMG bridges the gaps in clinical researcher abilities, allowing researchers at a low level of technical skill to analyze respiratory EMG.  
 
 
 # Statement of need
