### 2023-03-09 ef7303d
Update resurfemg/helper_functions.py
Co-authored-by: Walter Baccinelli <100706999+wbaccinelli@users.noreply.github.com>
