+++ resurfemg/helper_functions.py
@@ -1551,7 +1551,10 @@ def variability_maker(array, segment_size, method='variance'):
     """
     variability_values = []
     if method == 'variance':
-        variability_values = [np.var(array[i:i + segment_size]) for i in range(1, len(array) - segment_size)]
+        variability_values = [
+            np.var(array[i:i + segment_size])
+            for i in range(1, len(array) - segment_size)
+        ]
 
         values_out = np.array(variability_values)
 
