### 2023-03-09 716da6f
add walter suggestion and notebook
