+++ docs/resurfemg.rst
@@ -7,10 +7,12 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
+   resurfemg.cli
+   resurfemg.config
    resurfemg.converter_functions
    resurfemg.helper_functions
+   resurfemg.ml
    resurfemg.multi_lead_type
-   resurfemg.config
    resurfemg.tmsisdk_lite
 
 Module contents
