+++ docs/resurfemg.multi_lead_type.rst
@@ -1,5 +1,5 @@
 resurfemg.multi\_lead\_type module
-=====================================
+==================================
 
 .. automodule:: resurfemg.multi_lead_type
    :members:
