+++ docs/resurfemg.tmsisdk_lite.rst
@@ -1,5 +1,5 @@
 resurfemg.tmsisdk\_lite module
-==================================
+==============================
 
 .. automodule:: resurfemg.tmsisdk_lite
    :members:
