+++ docs/resurfemg.cli.rst
@@ -1,5 +1,5 @@
 resurfemg.cli module
-==================================
+====================
 
 .. automodule:: resurfemg.cli
    :members:
