### 2023-03-13 #5251 @editorialbot
[REVIEW]: ReSurfEMG: A Python library for preprocessing and analysis of respiratory EMG.
**Submitting author:** @drcandacemakedamoore  (<a href="http://orcid.org/0000-0003-1672-7565">Candace Makeda Moore</a>)
**Repository:** https://github.com/ReSurfEMG/ReSurfEMG
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
