### 2023-03-16 6bc022f
add csv to numpy file maker in cli
