### 2023-03-16 c706576
Update resurfemg/converter_functions.py
will try

Co-authored-by: Walter Baccinelli <100706999+wbaccinelli@users.noreply.github.com>
