+++ README.md
@@ -134,8 +134,24 @@ metadata is available.
    * Linux users can create their own environment by hand (use
      install_dev as in setup).
 
+Option B: In theory if you want to work, but never develop, as a conda user
+   with the stable version create an empty environment, and install
+   there: 
 
-  Option B: In theory if you want to work, but never develop, as a conda user
+
+   * Create a blank environment with python pinned to 3.8:
+
+     ```sh
+     conda create -n blank python=3.8
+     ```
+
+   * Install within the blank environment
+      ```sh
+        conda activate blank
+        conda install -c conda-forge -c resurfemg resurfemg jupyter ipympl
+        ```
+
+  Option B2: In theory if you want to work, but never develop, as a conda user
    with the stable version create an empty environment, and install
    there:
 
