+++ CHANGELOG.md
@@ -95,4 +95,5 @@
 
 	- Improved function for entropy in helper_functions module
 	- Variability over array function (variability_maker) in helper_functions module
+	- Additional command line functionality to convert certain csv files into files that can be processing in the ReSurfEMG-Dashboard.
 	
