### 2023-03-20 #5251 @AJQuinn
Hi @marcoghislieri @ixjlyons - its been around a week so I'd like to check in and make sure all is in place for you to get started on the review?

You can create a list of key points to assess for the review by commenting `@editorialbot generate my checklist` in this thread. Then the main review can take place over in `ReSurfEMG`'s issues board - don't forget to tag this thread in any issues that you open so we can keep track.

Let me know if you have face any issues or have any questions.

Cheers
