### 2023-03-28 36d34e5
Updated function for poly5 files Duiverman
