### 2023-04-03 b623aa9
update notebook guide
