+++ notebooks_guide.md
@@ -1,70 +1,29 @@
 # Notebooks guide :rainbow:
 
-You will see various notebooks you can use and adapt alongside our package. If you read the whole document we will describe each notebook (work in progress). For the impatient, here are some quick links :fast_forward: :
+ReSurfEMG contains various notebooks you can use and adapt alongside our package. The researcher_interface folder contains notebooks that are more or less 'standard' procedures. The open_work folder contains notebooks with more experimental work. This document is a work in progress overview of all notebooks. Below is a list providing a general overview of the notebooks and their purpose. 
 
-> * [You want to configure everything to run on your own data](#configuration)
-> * [You want to properly align your ventilator and EMG signal](#align)
-> * [You want to understand gating](#gating)
-> * [You want to see some basic analysis](#analysis)
-> * [You want to make array-snippets for ML](#snippets)
-> * [You want to see some ML in action](#mlaction)
+## **Researcher_ interface**
 
-Now a *comprehensive list* (under development, missing many notebooks as of 17/09/2022 :construction:): 
+> #### Basic EMG analysis
 
+>> This notebook provides [the basic EMG analysis](https://github.com/ReSurfEMG/ReSurfEMG/blob/main/researcher_interface/basic_emg_analysis.ipynb): picking a data file and sample, visualization, ECG removal with ICA and usage of the working_pipeline_pre_ml from helper_functions, resulting in calculation of length of breaths and absolute and percentage of total breath location of peaks. Area under the curve under construction.
 
+> #### Entropy
 
-> #### Configuration
-
->> We include a [notebook which explains how to configure your data paths](http://localhost:8888/notebooks/open_work/config_demo.ipynb).
-
+>> There are multiple entropy notebooks; this notebook provides [multiple approaches to use entropy for breath detection](https://github.com/ReSurfEMG/ReSurfEMG/blob/main/researcher_interface/entropy_eline_near_final.ipynb), using different thresholds and visualization on breath by breath basis, as well as a manual check of breath count. We can use it in various ways to automatically identify inhalation  [using multiple cut-offs](https://github.com/ReSurfEMG/ReSurfEMG/blob/main/researcher_interface/widgeted_entropy_updated_june.ipynb) or [a single cutt off](https://github.com/ReSurfEMG/ReSurfEMG/blob/main/researcher_interface/widgeted_entropy1.ipynb). Entropy is also used as part of some workflows to determine stegnth of respiratory efforts (listed in a seperate section).
 
 > #### Align
 
 >> Lag lead mismatch: You can either upsample the less sampled lead or downsample the more frequently sampled lead. For us only upsampling works as shown [in this lag-lead notebook](https://github.com/ReSurfEMG/ReSurfEMG/blob/main/researcher_interface/lead_lag_match_upsample.ipynb) but you can also try it the other way i.e. downsampling the fast signal [here](https://github.com/ReSurfEMG/ReSurfEMG/blob/main/researcher_interface/lag_lead_match.ipynb).
 
 
-> #### Array time matching
+> #### Machine learning 
 
[…]
