### 2023-04-04 0100c67
update notebook guide
