### 2023-04-04 d988b8d
add functions for new entropy
