+++ resurfemg/helper_functions.py
@@ -1599,3 +1599,192 @@ def variability_maker(
         variability_array = np.hstack((values_out, last_values_avg))
 
     return variability_array
+
+
+def rowwise_chebyshev(x, y):
+    return np.max(np.abs(x - y), axis=1)
+
+
+def delay_embedding(data, emb_dim, lag=1):
+    """
+    The following code is adapted from openly licensed code written by
+    Christopher Schölzel in his package nolds
+    (NOnLinear measures for Dynamical Systems).
+    It performs a time-delay embedding of a time series
+
+    :param data: array-like
+    :type data: array
+    :param emb_dim: the embedded dimension
+    :type emb_dim: int
+    :param lag: the lag between elements in the embedded vectors
+    :type lag: int
+
+    :returns: matrix_vectors
+    :rtype: ~nd.array
+    """
+    data = np.asarray(data)
+    min_len = (emb_dim - 1) * lag + 1
+    if len(data) < min_len:
+        msg = "cannot embed data of length {} with embedding dimension {} " \
+            + "and lag {}, minimum required length is {}"
+        raise ValueError(msg.format(len(data), emb_dim, lag, min_len))
+    m = len(data) - min_len + 1
+    indices = np.repeat([np.arange(emb_dim) * lag], m, axis=0)
+    indices += np.arange(m).reshape((m, 1))
+    matrix_vectors = data[indices]
+    return matrix_vectors
+
+
+def sampen(
+        data,
+        emb_dim=2,
+        tolerance=None,
+        dist=rowwise_chebyshev,
+        closed=False,
+        debug_plot=False,
+        debug_data=False,
+        plot_file=None
+):
+    """
+    The following code is adapted from openly licensed code written by
+    Christopher Schölzel in his package
+    nolds (NOnLinear measures for Dynamical Systems).
+    It computes the sample entropy of time sequence data.
+    Reference:
+        .. [se_1] J. S. Richman and J. R. Moorman, “Physiological time-series
+        analysis using approximate entropy and sample entropy,”
+        American Journal of Physiology-Heart and Circulatory Physiology,
+        vol. 278, no. 6, pp. H2039–H2049, 2000.
+
+    Args:
+        data (array-like of float):
+        input data
+    Kwargs:
+        emb_dim (int):
+        the embedding dimension (length of vectors to compare)
+        tolerance (float):
+        distance threshold for two template vectors to be considered equal
+        (default: 0.2 * std(data) at emb_dim = 2, corrected for
+        dimension effect for other values of emb_dim)
+        dist (function (2d-array, 1d-array) -> 1d-array):
+        distance function used to calculate the distance between template
+        vectors. Sampen is defined using ``rowwise_chebyshev``. You should only
+        use something else, if you are sure that you need it.
+        closed (boolean):
+        if True, will check for vector pairs whose distance is in the closed
+        interval [0, r] (less or equal to r), otherwise the open interval
+        [0, r) (less than r) will be used
+        debug_plot (boolean):
+        if True, a histogram of the individual distances for m and m+1
+        debug_data (boolean):
+        if True, debugging data will be returned alongside the result
+        plot_file (str):
+        if debug_plot is True and plot_file is not None, the plot will be saved
+        under the given file name instead of directly showing it through
+        ``plt.show()``
+    Returns:
+        float:
+        the sample entropy of the data (negative logarithm of ratio between
+        similar template vectors of length emb_dim + 1 and emb_dim)
+        [c_m, c_m1]:
[…]
