### 2023-04-04 1540b4b
Start work on optimizing sampen()
