### 2023-04-05 eaf9621
Fix indentation in sampen()
