### 2023-04-05 e431e41
lint entropy function docu
