### 2023-04-05 a4883a4
Optimized sampen() should be now close performance-wise to other entropies
