### 2023-04-06 5ae08c9
minor fixes, reset dims to 1
