+++ workshop.yml
@@ -1,33 +1,0 @@
-#  environment document
-
-name: you_workshop
-channels:
-  - conda-forge
-  #- resurfemg
-dependencies:
-  - python=3.8.6
-  - ipympl
-  - ipywidgets=7.7.0
-  - jupyter=1.0.0
-  - matplotlib=3.3.3
-  - mne=0.21.2
-  - notebook=6.4.10
-  - numexpr=2.7.3
-  - numpy=1.22.3
-  - pandas
-  - pip=20.2
-  - pycodestyle
-  - pyxdf
-  #- resurfemg
-  - scikit-learn=1.1.1
-  - scipy=1.7.1
-  - sphinx=4.5.0
-  - textdistance=4.2.1
-
-  - pip:
-
-    - resurfemg
-
-
-
-
