+++ .github/workflows/on-commit.yml
@@ -51,6 +51,7 @@ jobs:
       github.ref == 'refs/heads/drcandacemakedamoore/remove_sphinx' ||
       github.ref == 'refs/heads/wvxvw/sphinx-command' ||
       github.ref == 'refs/heads/drcandacemakedamoore/sphinx_and_nb' ||
+      github.ref == 'refs/heads/drcandacemakedamoore/expandmatrix' ||
       github.ref == 'refs/heads/wvxvw/documentation-fixes' ||
       github.ref == 'refs/heads/wvxvw/readthedocs-integration'
     steps:
