+++ .github/workflows/on-commit.yml
@@ -50,6 +50,7 @@ jobs:
       github.ref == 'refs/heads/main' ||
       github.ref == 'refs/heads/drcandacemakedamoore/remove_sphinx' ||
       github.ref == 'refs/heads/drcandacemakedamoore/sphinx_and_nb' ||
+      github.ref == 'refs/heads/drcandacemakedamoore/expandmatrix' ||
       github.ref == 'refs/heads/wvxvw/documentation-fixes' ||
       github.ref == 'refs/heads/wvxvw/readthedocs-integration'
     steps:
