### 2023-05-03 e7df80b
Adopt former Sphinx setuptools integration
