+++ .github/workflows/on-commit.yml
@@ -49,6 +49,7 @@ jobs:
     if: |
       github.ref == 'refs/heads/main' ||
       github.ref == 'refs/heads/drcandacemakedamoore/remove_sphinx' ||
+      github.ref == 'refs/heads/wvxvw/sphinx-command' ||
       github.ref == 'refs/heads/drcandacemakedamoore/sphinx_and_nb' ||
       github.ref == 'refs/heads/wvxvw/documentation-fixes' ||
       github.ref == 'refs/heads/wvxvw/readthedocs-integration'
