### 2023-05-03 e76bcbd
Adopt former Sphinx setuptools integration
