+++ CHANGELOG.md
@@ -89,12 +89,12 @@
 
 ## Release 0.0.9
 
-### Not yet added
+### Added
 
 * Ninth version of this Python project, containing (added to this version):
 
 	- Improved function for entropy in helper_functions module
 	- Variability over array function (variability_maker) in helper_functions module
 	- Additional command line functionality to convert certain csv files into files that can be processing in the ReSurfEMG-Dashboard.
-	- Expanded testing matrix to include Python 3.8 and 3.9
+	- Expanded testing matrix to include Python 3.8 and 3.9 
 	
