+++ CHANGELOG.md
@@ -97,4 +97,10 @@
 	- Variability over array function (variability_maker) in helper_functions module
 	- Additional command line functionality to convert certain csv files into files that can be processing in the ReSurfEMG-Dashboard.
 	- Expanded testing matrix to include Python 3.8 and 3.9 
+
+## Release 0.0.10
+
+### Not yet added
+
+* Tenth version: 
 	
