### 2023-05-24 #5251 @AJQuinn
Hello @behinger, @ajbaird and @peterakirk!

This is a review thread for a Python package "ReSurfEMG: A Python library for preprocessing and analysis of respiratory EMG" at the Journal of Open Source Software. More details in the links in the top comment.

Review has started - but we need to find another reviewer to join us at short notice. Would any of you be willing to contribute a review for this submission for JOSS? We'd ideally like to find someone who could get started soon as the review has been going for a while now.

JOSS carries out checklist-driven reviews here in GitHub issues. Reviews follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
