### 2023-06-20 #5251 @ixjlyons
Hi all, my sincere apologies for taking so long to get this review in. This one really got away from me. Anyway, here's my review.

----

## Overall Comments

I think the purpose of the software is clear and described well in the paper. Having worked for some time in a related field, I can attest to lack of clarity in algorithm implementations and parameters. I have some overall feedback that I think could improve the repository, but otherwise it appears to be in good shape.

The main constructive feedback I have for ReSurfEMG is to spruce up the analysis notebooks a bit more, perhaps with some additional narrative explanations weaved around the code cells. I think most of them already do this pretty well, but some start to get heavy with code and sparser explanations toward the end. You might also consider changing the usage of widgets in the notebooks to update analysis/plots as user choices change (datapoint, number of samples, etc.) instead of having to go back to the widget and then run subsequent cells. My usual approach to this is to choose a default set of inputs, write out the analysis steps via functions with intermediate evaluations to walk through what's being done, then hook up the function(s) to widgets for interactivity. I consider these recommendations/suggestions, not mandatory for acceptance.

## Paper

My only comments on the paper are minor:

- The second sentence of the Summary is somewhat awkward -- I think there's some redundancy in "to meet challenges in interpretation of the signals," so perhaps that could be removed.
- Remove the comma on line 50 ("...complex EMG data sets remains very limited...")

## Documentation

Installation notes:

- Creating a conda environment from the `environment.yml` file ran for a very long time resolving, so I didn't end up fully testing this installation method. It's not obvious what would cause this given most of the dependencies are pinned to versions that presumably were all compatible at some point, but it might be worth trying again to see if you can reproduce the lengthy dependency resolution.
- Developer's setup:
    - There's a comment that Windows users activate the environment with `.venv/bin/activate`, but I think this should be more like `.venv/Scripts/activate`
    - Initial (pip) install failed the message below. I'm not sure what the cause is, but running `pip install jupyter` afterward seems to have resolved it.

```
Searching for notebook
Reading https://pypi.org/simple/notebook/
Downloading https://files.pythonhosted.org/packages/67/ae/fd5ce79fb604fb4a9bc582bfa24dd3b27d85cdcee5d646abb3f905deebfb/notebook-7.0.0b0-py3-none-any.whl#sha256=a73f295ad0e728d01af316ed04cadb39913da1e77647faf57a3c4f6b3844a303
Best match: notebook 7.0.0b0
Processing notebook-7.0.0b0-py3-none-any.whl
Installing notebook-7.0.0b0-py3-none-any.whl to /home/krlyons/tmp/ReSurfEMG/.venv/lib/python3.8/site-packages
error: [Errno 13] Permission denied: '/home/krlyons/tmp/ReSurfEMG/.venv/lib/python3.8/site-packages/notebook-7.0.0b0-py3.8.egg/notebook-7.0.0b0.data/data/share' -> '/home/krlyons/tmp/ReSurfEMG/.venv/lib/python3.8/site-packages/notebook-7.0.0b0-py3.8.egg/./share'
```

Due to the notes above, I'll leave the installation box unchecked for now, but I was ultimately able to get the package installed to run the analysis notebooks and try out ResurfEMG-dashboard.

## Data

- It might be nice to consider uploading the sample dataset somewhere it can more easily be retrieved (without Docker), e.g. Dryad, Zenodo, OSF.
- Perhaps also a walkthrough of generating a synthetic dataset sufficient to get started on a notebook or two would be good.
- It might also be advisable to support unit testing without retrieving an external dataset, either by including a small subset necessary for testing, relying on synthetic data, or maybe marking certain tests as skippable if data aren't present. [Git LFS](https://git-lfs.com/) might be worth looking into.
- Once the data has been copied out of the docker image, it's not fully clear how to create a config that points to it for the notebooks. I was able to figure it out, but it could be spelled out a little more explicitly.

## Code

The Python package has a somewhat odd structure/organization. I think the main issue is "helper functions" isn't particularly descriptive and contains the bulk of the code. Maybe it could be split up a bit for clarity and/or more manageable expansion as new methods are added (visualzation, filtering, feature extraction, etc.). Just a suggestion.
