### 2023-06-23 #5251 @ajbaird
Hi All, Here are my thoughts

Overview
-----------

The document is well constructed and checks all of the boxes needed for a JOSS publication.  Statement of need is clear state of the field and summery are well written. I do wish there was a section on use case, or context of use that could guide a user around API and data sets available.

I found a large lack of clarity around the repository documentation, especially regarding datasets that can be used to test this software, how they can be leveraged and a description of the output targets.

Paper
----------

- line 25, delete "much"
- line 26 define "filter settings"
- line 39 delete "aside from lacking transparency in clinical scientific literature" seems like a strange jab without adding to the discussion
- line 61 is a little confusing, maybe re-write especially the part after the "...remove noise, and .... " are the signals the cardiac signals or are they the noise, the signal is again referenced in the next paragraph and is similarly confusing.

Documentation
----------------------
install
======
The notes here weren't complete for all platforms (non conda environments) and had an issue with tk as a dep. Here are my notes:
- virtual environment on a ubuntu virtual machines dev install worked well
- Shouldn't be burried at the end of the documentation, it also shouldn't be called "developers setup" that is quite confusing
- Getting started should include installations steps for each architecture supported, currently only conda
- No tutorial on how to use synthetic data to run a simulation to verify functionality
- tkinter not included on the dependencies of the virtual environment
- There is an issue with the tkinter dependencies, its important to spell out the dep in your venv configuration file
	- I was able to work around this by installing tkinter via: apt-get install python3-tk, it didn't resolve the issue by using pip

Proof of context
==============
I aimed to work my way through the jupyter notebook titled basic_emg_analysis using the zenodo synthetic data set. I had the following issues regarding this process and was not able to generate the figures at the end of the notebook. Here are my notes:

 why is the directory information for the data at least not set to some default value, it is very confusing setting up new directories and configure files that restrict even running the most basic example without
	- Can you just set them to a default location, then in your readme tell the user to download the synthetic data set and unzip it to the following location, then you can describe to them how to change that location as needed
- your example json formatting doesn't work across architectures, double quotes are needed (in the config.json file)
- its required that you create the /data/root_emg directory, this should be created but empty in the project OR should be auto created to a default location if the json config file isn't provided
- There is synthetic data in a directory when you clone the project, can we edit the basic_emg_analysis file to use this data?
- testing isn't defined in the readme, only described for a docker container
- In the notebook there is an assumption of how the emg files look (''**/*.Poly5') but this isn't consistent with the synthetic data provided in zenodo
- Data is in .npy files not sure about the assumed data structure here
- Needed to change the cell to be emg_pattern = os.path.join(root_emg_directory, 'data/*.npy')
- this is after I unzipped and moved the synthetic data into the appropriate root directory
- The polyreader class doesn't work with synthetic data file, I get the following error (with hard coded indexing on the date/time routine)
- ValueError: year 40070 is out of range, in the line item self.start_time=datetime.datetime(header_data[8], header_data[9], header_data[10]), ect...
- This is clearly assuming the data is in a certain format (and not the one that is presented in the synthetic data), this is a huge issue and needs patching

Conclusions
---------------
Considering these issues in getting a viable proof of context and use, I feel like there are edits and changes needed to the repo documentation in order to approve publication of this paper. Please address to improve documentation.
