+++ .github/workflows/on-commit.yml
@@ -6,7 +6,7 @@ jobs:
     strategy:
       matrix:
         os: ['ubuntu-latest']
-        python-version: ['3.8', '3.9']
+        python-version: ['3.9']
     steps:
       - uses: actions/checkout@v3
         with:
@@ -39,7 +39,7 @@ jobs:
       - uses: s-weigand/setup-conda@v1
         with:
           conda-channels: conda-forge
-          python-version: 3.8
+          python-version: 3.9
       - run: python3 -m venv .venv
       - run: .venv/bin/python -m pip install wheel
       - run: .venv/bin/python setup.py lint
@@ -62,7 +62,7 @@ jobs:
       - uses: s-weigand/setup-conda@v1
         with:
           conda-channels: conda-forge
-          python-version: 3.8
+          python-version: 3.9
       - run: python3 -m venv .venv
       - run: .venv/bin/python -m pip install wheel sphinx
       - run: .venv/bin/python setup.py install
