+++ .helpers/Dockerfile
@@ -1,4 +1,4 @@
-from python:3.8
+from python:3.9
 
 run mkdir -p /ReSurfEMG/tests/not_pushed/
 add ./not_pushed /ReSurfEMG/tests/not_pushed/
