### 2023-06-24 #5251 @drcandacemakedamoore
Thank you for your thoughtful reviews @ajbaird  and @ixjlyons .  There are many changes we will work on over the summer. I would note that there is synthetic data at https://github.com/ReSurfEMG/synthetic_data which has been archived on Zenodo. As the data is in  .npy format we will better instructions in cases where .npy can not be used directly i.e. where we begin with a Poly5 file. There are many converter functions for different lab groups on the repository in converter_functions.py; but a converter function to a Poly5 is a bit of an unnecessary step backwards. We really appreciate your thoughtful and thorough reviews, and will notify you as we update the library and paper.
