### 2023-06-27 bdc8983
Some sandbox notebooks for assessing curve quality
