### 2023-07-08 7be2ece
Upgrade environment.yml to Python 3.9
