+++ environment.yml
@@ -1,15 +1,15 @@
 #  environment document
 
-name: julytry2
+name: julytry3
 channels:
   - conda-forge
 dependencies:
-  - python=3.8.6
+  - python=3.9
   - ipympl
   - ipywidgets=7.7.0
   - jupyter=1.0.0
   - matplotlib=3.3.3
-  - mne=0.21.2
+  - mne=0.23.4
   - notebook=6.4.10
   - numexpr=2.7.3
   - numpy=1.22.3
