+++ paper/paper.md
@@ -37,18 +37,18 @@ bibliography: paper.bib
 
 
 ReSurfEMG is an open source collaborative Python library for analysis of respiratory surface electromyography (EMG).
- Analysis and interpretation of respiratory surface EMG requires expertise of both clinicians and technical specialists including software engineers, to meet challenges in interpretation of the signals, for example identification of the onset and offset of respiratory effort. This neccesary cooperation of clinicians and technical staff implies that respiratory EMG data must be considered by groups of professionals who differ profoundly in computational skill levels. Therefore, our package implements different processing and analytic pipelines for researchers to investigate, through several interfaces. The package not only allows for pre-processing and analysis with either command line or Jupyter notebooks, it supports a graphic user interface package which allows code-free interaction with the data, the [ReSurfEMG-dashboard](https://github.com/ReSurfEMG/ReSurfEMG-dashboard). The interface for ReSurfEMG algorithms through Jupyter notebooks allows researchers to run preset experiments, potentially extend them and record the results in csv files automatically.
-ReSurfEMG code allows for analysis of various parameters in respiratory surface EMG from simple ones as peak amplitude to area under curve and the potential for more complex ones such as entropy. These characteristics can be used in machine learning models, for which there is also code in the notebooks to demonstrate. The current state of research in the field includes much scientific work without any published code and therefore it is currently difficult to compare outcome parameters in (clinical) research, as filter settings may differ and algorithms for feature extraction may differ. This library enables communication towards reproducible research. 
+ Analysis and interpretation of respiratory surface EMG requires expertise of both clinicians and technical specialists including software engineers, for example to identify the onset and offset of respiratory effort. This neccesary cooperation of clinicians and technical staff implies that respiratory EMG data must be considered by groups of professionals who differ profoundly in computational skill levels. Therefore, our package implements different processing and analytic pipelines for researchers to investigate, through several interfaces. The package not only allows for pre-processing and analysis with either command line or Jupyter notebooks, it supports a graphic user interface package which allows code-free interaction with the data, the [ReSurfEMG-dashboard](https://github.com/ReSurfEMG/ReSurfEMG-dashboard). The interface for ReSurfEMG algorithms through Jupyter notebooks allows researchers to run preset experiments, potentially extend them and record the results in csv files automatically.
+ReSurfEMG code allows for analysis of various parameters in respiratory surface EMG from simple ones as peak amplitude to area under curve and the potential for more complex ones such as entropy. These characteristics can be used in machine learning models, for which there is also code in the notebooks to demonstrate. The current state of research in the field includes scientific work without any published code and therefore it is currently difficult to compare outcome parameters in (clinical) research. Filter settings as cutoff frequencies and window lengths to create a respiratory tracing may differ, as well as the used algorithms for feature extraction. This library enables communication towards reproducible research. 
 
 
 # State of the field
 
[…]
