### 2023-07-15 6fc3a18
checking array and other factors in new mne
