### 2023-07-21 7231ac5
removing config from cli
