+++ resurfemg/cli.py
@@ -173,16 +173,15 @@ def main(argv):
     """
     parser = make_parser()
     parsed = parser.parse_args()
-    config = Config(parsed.config)
 
     if parsed.action == 'acquire':
         try:
 
             preprocess(
-                config.get_directory('data', parsed.input),
+                parsed.input,
                 parsed.lead or [0, 2],  # list of chosen leads
                 parsed.preprocessing,
-                config.get_directory('preprocessed', parsed.output),
+                parsed.output,
                 parsed.force,
             )
         except Exception as e:
@@ -193,8 +192,8 @@ def main(argv):
         try:
 
             save_j_as_np(
-                config.get_directory('data', parsed.input),
-                config.get_directory('made', parsed.output),
+                parsed.input,
+                parsed.output,
             )
         except Exception as e:
             logging.exception(e)
@@ -204,9 +203,9 @@ def main(argv):
         try:
 
             make_realistic_syn_emg_cli(
-                config.get_directory('data', parsed.input),
+                parsed.input,
                 parsed.number,
-                config.get_directory('made', parsed.output),
+                parsed.output,
             )
         except Exception as e:
             logging.exception(e)
@@ -215,9 +214,9 @@ def main(argv):
     if parsed.action == 'ml':
         try:
             applu_model(
-                config.get_directory('data', parsed.input),
+                parsed.input,
                 parsed.model,
-                config.get_directory('preprocessed', parsed.output),
+                parsed.output,
                 parsed.features,
             )
         except Exception as e:
