### 2023-07-21 68ce798
added testing and new notebook to readme file
