### 2023-07-31 7059cd3
Revert "removing config from cli"
This reverts commit dc0ff066e1e1e4ce6d7f5596fa10db1692a4bf0d.
