### 2023-07-31 1b61752
use config if paths are empty
