### 2023-08-03 ReSurfEMG/ReSurfEMG#252 @drcandacemakedamoore
communicate to JOSS
- we have a new environment approach
- as well as updated synthetics data, and some updated notebooks
- although we are finishing some other issues, we hope we can find out if the current environment installation is satifactory for you
