### 2023-08-05 2bd289a
add etc folder and more test
