+++ README.md
@@ -198,7 +198,7 @@ that versions 3.8 or 3.9 (depending on the release) are supported.
 
 ### Installation for all supported platforms
 
-Installation with Anaconda/conda and/or mamba are covered in [getting started](#Getting Started). If you wish to install with pip.... is Here we need to add pip instructions.
+Installation with Anaconda/conda and/or mamba are covered in [getting started](#Getting-Started). If you wish to install with pip.... is Here we need to add pip instructions.
 
 
 ## Developer's setup
