### 2023-08-06 a0d0d10
perfection
