### 2023-08-11 #5251 @drcandacemakedamoore
@Kevin-Mattheus-Moerman , @ixjlyons and @ajbaird Thank you all for your thoughtful comments and review. We have now made several changes to the repository. In fact we made almost all the changes suggested except splitting up helper functions. We felt this part of the project should be worked on with the main users, so it is still ongoing. I believe I have fixed the tkinter issue,  and my installation instructions are more clear and inclusive.  I would love to hear if this is now working for everyone.
