### 2023-08-14 #5251 @AJQuinn
Hi @drcandacemakedamoore - thanks for the response to feedback. I can see lots of changes on the repo but I'm afraid I'm struggling to match them up to bits of specific feedback.

To make the next step simpler, would you mind commenting with a quick summary of specific PRs/commits that relate specifically to the JOSS review? or better still, link this thread into the comments of an active issues/PR. They should then appear here for reference.

@Kevin-Mattheus-Moerman , @ixjlyons and @ajbaird - please can you try and get to this in the next two weeks if at all possible. Please let me know if you'll need more time.
