### 2023-08-14 ReSurfEMG/ReSurfEMG#252 @drcandacemakedamoore
They need a table like this:

|                             | Suggestion  | Commit   NUMBER   | file and area       |
|:---------------------------:|:---------:|:---------:|:---------:|
|  | Do A | Unknown   | Unknown   |
| | dO b | # | Supported |
