### 2023-08-14 eb4b8b3
[cr] Reworded paragraph about developers roles
