### 2023-08-15 #5251 @drcandacemakedamoore
[Hot linked version of correction table] (https://github.com/ReSurfEMG/ReSurfEMG/issues/252#issuecomment-1677346866) being built
