### 2023-08-15 12061b0
clarify install_dev further
