### 2023-08-15 #5251 @drcandacemakedamoore
I think I have responded to almost all comments by providing links to the PR and commit number, or explaining our ideas on the issue here: (https://github.com/ReSurfEMG/ReSurfEMG/issues/252#issuecomment-1677346866). We have to work with the main end users of the library at the university medical centers to split up the helper  functions  module in a way that makes sense, and this will take time.
