### 2023-08-15 #5251 @ajbaird
hi @drcandacemakedamoore thank you so much! I thought that the repo was super duper close to checking all the boxes for me, just with a better readme and instructions on linux installation. I'll be able to look over it in the coming two weeks! Thanks again for all the hard work!
