### 2023-09-04 #5251 @ixjlyons
> We made almost all the changes suggested except splitting up helper functions. We felt this part of the project should be worked on with the main users, so it is still ongoing.

Completely reasonable. It was meant as a suggestion, but I'm happy to hear you're working on it. I think it'll make the code a little easier to start working with, but good choice taking care of existing users.

----

I've reviewed the changes made and it all looks good to me. I went through the installation instructions and ran through a couple notebooks again to check, so I've checked off everything in my checklist.
