### 2023-09-22 #5251 @ajbaird
Hi All!

Was just able to pull the repo with all the changes you've commmited and I'm really impressed! you've absolutely hit all the issues I had last time around. Super easy to begin working on the notebooks on a linux platform. I really commend you for all the effort you put into it. Thank you. I know it was a lot of asks and everyone is so busy these days :)

@AJQuinn Paper is good to go from my point of view. Excellent tools, excellent documentation. Really well done!

Thanks again @drcandacemakedamoore this repo will be a treat for a lot of people to play around with!

Austin
