### 2023-09-27 #5251 @drcandacemakedamoore
Thank you guys who reviewed ( @ixjlyons  and @ajbaird )! Your comments were really useful in terms of making the package more broadly usable, and the user side of the team is and will continue refactoring and adding a bit more code based on some of them. Please let us know if there is anything missing to go to the next steps in the publication process @AJQuinn .
