### 2023-10-04 ReSurfEMG/ReSurfEMG#263 @AJQuinn
Missing DOI for JOSS publication
Hi @drcandacemakedamoore - we'll need to fix this missing DOI issue prior to accepting the paper at JOSS.

The missing DOI is listed here: https://github.com/openjournals/joss-reviews/issues/5251#issuecomment-1747323849

and seems to point to the reference here: https://github.com/ReSurfEMG/ReSurfEMG/blob/36876d604a6ae60016cfec95c3fed7142688054b/paper/paper.bib#L77

I suspect the issue is that there isn't an explicit DOI line in the bibtex for that reference. Could you try adding something like

`  doi = {10.1007/978-3-642-34546-3_71},
`
to the bibtex and rerunning the compile pdf and check reference commands on the JOSS thread?

https://github.com/openjournals/joss-reviews/issues/5251
