### 2023-10-04 6af6df1
Update DOI in Silva et al reference in paper.bib
