+++ paper/paper.bib
@@ -89,7 +89,8 @@ publisher={Springer Berlin Heidelberg},
 address={Berlin, Heidelberg},
 pages={437-441},
 abstract={Biosignals have been extensively used within the field of Human- Computer Interaction (HCI) for the development of brain and neuronal computer interaction systems, in particular, targeting augmentative and alternative communication. Electromyography is a commonly used technique in this context, however, conventional signal acquisition procedures require the use of conductive paste or pre-gelled electrodes on the interface with the skin, thus introducing usability constraints and recurrent costs associated with the daily use. Our work is focused on the comparison between standard pre-gelled Ag/AgCl electrodes, dry Ag/AgCl electrodes and conductive fabrics, as a way of improving the usability of electromyographic interfaces for HCI. Experimental results have shown that dry and gelled electrodes perform similarly, thus applicability of alternative electrode materials, and further extending the toolbox of the rehabilitation engineer.},
-isbn={978-3-642-34546-3}
+isbn={978-3-642-34546-3},
+doi={10.1007/978-3-642-34546-3_71}
 }
 
 @article{londral2013wireless,
