+++ paper/paper.bib
@@ -145,6 +145,7 @@ journal = {Journal of Open Source Software} }
   year={2017}
 }
 
+
 @Misc{Carreiras2015biosppy,
   author = {Carlos Carreiras and Ana Priscila Alves and Andr\'{e} Louren\c{c}o and Filipe Canento and Hugo Silva and Ana Fred and others},
   title = {{BioSPPy}: Biosignal Processing in {Python}},
