### 2023-10-06 #5251 @AJQuinn
I'll have a read through the paper and make any additional suggestions in a PR (will be minor if anything at all).

@drcandacemakedamoore could you double check authors and affiliations etc as well? Once that's confirmed we can work through the list above.
