### 2023-10-06 36361ec
Suggested types/fix-ups for JOSS paper
