### 2023-10-06 #5251 @AJQuinn
Ok - I've put a few suggested paper tweaks in the PR above, please review and action these as you prefer.

Next step is to work through the 'Additional Author Tasks After Review is Complete' steps in the checklist above. Let me know if anything issues come up.
