+++ resurfemg/config/config.py
@@ -1,0 +1,323 @@
+"""
+Copyright 2022 Netherlands eScience Center and Twente University.
+Licensed under the Apache License, version 2.0. See LICENSE for details.
+
+This file contains one method to let the user configure
+all paths for data instead of hard-coding them, as well
+as methods to check data integrity, and create synthetic data.
+The data integrity can be checked because this file contains
+hash functions to track data. Synthetic data can be made
+with several methods.
+"""
+
+import json
+import logging
+import os
+import textwrap
+import hashlib
+import glob
+import math
+import pandas as pd
+import numpy as np
+import scipy
+from scipy import signal
+import random
+
+
+class Config:
+    """
+    This class allows configuration on the home computer
+    or remote workspace, of a file setup for data,
+    which is then processed into a variable. Essentially
+    by setting up and modifying a .json file in the appropriate directory
+    users can avoid the need for any hardcoded paths to data.
+    """
+
+    default_locations = (
+        './config.json',
+        os.path.expanduser('~/.resurfemg/config.json'),
+        '/etc/resurfemg/config.json',
+    )
+
+    default_layout = {
+        'root_emg_directory': '{}',
+        'preprocessed': '{}/preprocessed',
+        'models': '{}/models',
+        'output': '{}/output',
+    }
+
+    required_directories = ['root_emg_directory']
+
+    def __init__(self, location=None):
+        self._raw = None
+        self._loaded = None
+        self.load(location)
+        self.validate()
+
+    def usage(self):
+        """
+        This is essentally a corrective error message if the computer
+        does not have paths configured or files made so that
+        the data paths of config.json can be used
+        """
+        return textwrap.dedent(
+            '''
+            Cannot load config.
+
+            Please create a file in either one of the locations
+            listed below:
+            {}
+
+            With the contents that specifies at least the root
+            directory as follows:
+
+            {{
+                "root_emg_directory": "/path/to/storage"
+            }}
+
+            The default directory layout is expected to be based on the above
+            and adding subdirectories.
+
+            You can override any individual directory (or subdirectory)
+            by specifying it in the config.json file.
+
+            "root_emg_directory" is expected to exist.
+            The "models" and "preprocessed" directories need not
+            exist.  They will be created if missing.
+            '''
+        ).format('\n'.join(self.default_locations))
+
+    def load(self, location):
+        locations = (
+            [location] if location is not None else self.default_locations
+        )
+
+        for p in locations:
+            try:
+                with open(p) as f:
+                    self._raw = json.load(f)
+                    break
+            except Exception as e:
+                logging.info('Failed to load %s: %s', p, e)
+        else:
+            raise ValueError(self.usage())
+
+        root = self._raw.get('root_emg_directory')
+        self._loaded = dict(self._raw)
+        if root is None:
+            required = dict(self.default_layout)
+            del required['root_emg_directory']
+            for directory in required.keys():
+                if directory not in self._raw:
+                    raise ValueError(self.usage())
+            # User specified all concrete directories.  Nothing for us to
+            # do here.
[…]
