### 2023-10-10 e5c9bfe
modify test naming convention in setup.py
