+++ resurfemg/helper_functions.py
@@ -78,232 +78,232 @@ class Range(namedtuple('RangeBase', 'start,end')):
         return slice(*map(int, self))   # maps whole tuple set
 
 
-def emg_bandpass_butter(data_emg, low_pass, high_pass):
-    """The parameter taken in here is the Poly5 file. Output is
-    the EMG after a bandpass as made here.
+# def emg_bandpass_butter(data_emg, low_pass, high_pass):
+#     """The parameter taken in here is the Poly5 file. Output is
+#     the EMG after a bandpass as made here.
 
-    :param data_emg: Poly5 file with the samples to work over
-    :type data_emg: ~TMSiSDK.file_readers.Poly5Reader
-    :param low_pass: The number to cut off :code:`frequenciesabove`
-    :type low_pass: int
-    :param high_pass: The number to cut off :code:`frequenceisbelow`
-    :type high_pass: int
+#     :param data_emg: Poly5 file with the samples to work over
+#     :type data_emg: ~TMSiSDK.file_readers.Poly5Reader
+#     :param low_pass: The number to cut off :code:`frequenciesabove`
+#     :type low_pass: int
+#     :param high_pass: The number to cut off :code:`frequenceisbelow`
+#     :type high_pass: int
 
-    :returns: The bandpass filtered EMG sample data
-    :rtype: ~numpy.ndarray
-    """
-    sos = signal.butter(
-        3,
-        [low_pass, high_pass],
-        'bandpass',
-        fs=data_emg.sample_rate,
-        output='sos',
-    )
-    # sos (output parameter) is second order section  -> "stabilizes" ?
-    emg_filtered = signal.sosfiltfilt(sos, data_emg.samples)
-    return emg_filtered
-
-
-def emg_bandpass_butter_sample(
-    data_emg_samp,
-    low_pass,
-    high_pass,
-    sample_rate,
-    output='sos'
-):
-    """The paramemter taken in here is the Poly5 file.  Output is the
-    EMG after a bandpass as made here.
-
-    :param data_emg_samp: The array in the Poly5 or other sample
-    :type data_emg_samp: ~numpy.ndarray
-    :param low_pass: The number to cut off :code:`frequenciesabove`
-    :type low_pass: int
-    :param high_pass: The number to cut off :code:`frequenceisbelow`
-    :type high_pass: int
-    :param sample_rate: The number of samples per second i.e. Hertz
-    :type sample_rate: int
-    :param output: The type of sampling stabilizor
-    :type high_pass: str
-
-    :returns: The bandpass filtered EMG sample data
-    :rtype: ~numpy.ndarray
-    """
-    sos = signal.butter(
-        3,
-        [low_pass, high_pass],
-        'bandpass',
-        fs=sample_rate,
-        output='sos',
-    )
-    # sos (output parameter)is second order section  -> "stabilizes" ?
-    emg_filtered = signal.sosfiltfilt(sos, data_emg_samp)
-    return emg_filtered
+#     :returns: The bandpass filtered EMG sample data
+#     :rtype: ~numpy.ndarray
+#     """
+#     sos = signal.butter(
+#         3,
+#         [low_pass, high_pass],
+#         'bandpass',
+#         fs=data_emg.sample_rate,
+#         output='sos',
+#     )
+#     # sos (output parameter) is second order section  -> "stabilizes" ?
+#     emg_filtered = signal.sosfiltfilt(sos, data_emg.samples)
+#     return emg_filtered
 
 
-def bad_end_cutter(data_emg, percent_to_cut=7, tolerance_percent=10):
-    """This algorithm takes the end off of EMGs where the end is
-    radically altered, or if not radically altered cuts the last 10
[…]
