+++ tests/preprocessing_test.py
@@ -1,0 +1,232 @@
+#sanity tests for the preprocessing functions, including filtering, ecg removal and envelope calculation
+
+
+import unittest
+import os
+import scipy
+import numpy as np
+from resurfemg.helper_functions import emg_bandpass_butter
+from resurfemg.helper_functions import emg_bandpass_butter_sample
+from resurfemg.helper_functions import bad_end_cutter
+from resurfemg.helper_functions import bad_end_cutter_better
+from resurfemg.helper_functions import bad_end_cutter_for_samples
+from resurfemg.helper_functions import notch_filter
+from resurfemg.helper_functions import emg_lowpass_butter
+from resurfemg.tmsisdk_lite import Poly5Reader
+from resurfemg.helper_functions import compute_ICA_two_comp
+from resurfemg.helper_functions import compute_ICA_two_comp_multi
+from resurfemg.helper_functions import pick_lowest_correlation_array
+from resurfemg.helper_functions import pick_more_peaks_array
+from resurfemg.helper_functions import gating
+from resurfemg.helper_functions import find_peaks_in_ecg_signal
+from resurfemg.multi_lead_type import pick_highest_correlation_array_multi
+from resurfemg.helper_functions import naive_rolling_rms
+from resurfemg.helper_functions import vect_naive_rolling_rms
+from resurfemg.helper_functions import smooth_for_baseline
+from resurfemg.helper_functions import smooth_for_baseline_with_overlay
+
+sample_emg = os.path.join(
+    os.path.abspath(os.path.dirname(os.path.dirname(__file__))),
+    'not_pushed',
+    'Test_lung_data',
+    '2022-05-13_11-51-04',
+    '002',
+    'EMG_recording.Poly5',
+)
+
+class TestFilteringMethods(unittest.TestCase):
+
+    def test_emg_band_pass_butter(self):
+        sample_read= Poly5Reader(sample_emg)
+        sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
+        self.assertEqual(
+            (len(sample_emg_filtered[0])),
+            len(sample_read.samples[0]) ,
+        )
+    def test_emg_band_pass_butter_sample(self):
+        sample_read= Poly5Reader(sample_emg)
+        sample_emg_filtered = emg_bandpass_butter_sample(sample_read.samples, 1, 10, 2048)
+        self.assertEqual(
+            (len(sample_emg_filtered[0])),
+            len(sample_read.samples[0]) ,
+        )
+    
+    def test_emg_lowpass_butter(self):
+        sample_read= Poly5Reader(sample_emg)
+        sample_emg_filtered = emg_lowpass_butter(sample_read.samples, 5, 2048)
+        self.assertEqual(
+            (len(sample_emg_filtered[0])),
+            len(sample_read.samples[0]) ,
+        )
+
+    def test_notch_filter(self):
+        sample_read= Poly5Reader(sample_emg)
+        sample_emg_filtered = notch_filter(sample_read.samples, 2048, 80,2)
+        self.assertEqual(
+            (len(sample_emg_filtered[0])),
+            len(sample_read.samples[0]) ,
+        )
+    def test_naive_rolling_rms(self):
+        sample_read= Poly5Reader(sample_emg)
+        sample_emg_filtered = naive_rolling_rms(sample_read.samples[0], 10)
+        self.assertNotEqual(
+            (len(sample_emg_filtered)),
+            len(sample_read.samples[0]) ,
+        )
+    
+    def test_vect_naive_rolling_rms(self):
[…]
