+++ tests/preprocessing_test.py
@@ -1,29 +1,27 @@
-#sanity tests for the preprocessing functions, including filtering, ecg removal and envelope calculation
+#sanity tests for the preprocessing functions, including filtering,
+#ecg removal and envelope calculation
 
 
 import unittest
 import os
 import scipy
 import numpy as np
-from resurfemg.helper_functions import emg_bandpass_butter
-from resurfemg.helper_functions import emg_bandpass_butter_sample
-from resurfemg.helper_functions import bad_end_cutter
-from resurfemg.helper_functions import bad_end_cutter_better
-from resurfemg.helper_functions import bad_end_cutter_for_samples
-from resurfemg.helper_functions import notch_filter
-from resurfemg.helper_functions import emg_lowpass_butter
-from resurfemg.tmsisdk_lite import Poly5Reader
-from resurfemg.helper_functions import compute_ICA_two_comp
-from resurfemg.helper_functions import compute_ICA_two_comp_multi
-from resurfemg.helper_functions import pick_lowest_correlation_array
-from resurfemg.helper_functions import pick_more_peaks_array
-from resurfemg.helper_functions import gating
-from resurfemg.helper_functions import find_peaks_in_ecg_signal
+from resurfemg.preprocessing.filtering import emg_bandpass_butter
+from resurfemg.preprocessing.filtering import emg_bandpass_butter_sample
+from resurfemg.preprocessing.filtering import bad_end_cutter
+from resurfemg.preprocessing.filtering import bad_end_cutter_better
+from resurfemg.preprocessing.filtering import bad_end_cutter_for_samples
+from resurfemg.preprocessing.filtering import notch_filter
+from resurfemg.preprocessing.filtering import emg_lowpass_butter
+from resurfemg.data_connector.tmsisdk_lite import Poly5Reader
+from resurfemg.preprocessing.ecg_removal import compute_ica_two_comp
+from resurfemg.preprocessing.ecg_removal import compute_ica_two_comp_multi
+from resurfemg.preprocessing.ecg_removal import pick_lowest_correlation_array
+from resurfemg.preprocessing.ecg_removal import pick_more_peaks_array
+from resurfemg.preprocessing.ecg_removal import gating
 from resurfemg.multi_lead_type import pick_highest_correlation_array_multi
 from resurfemg.helper_functions import naive_rolling_rms
 from resurfemg.helper_functions import vect_naive_rolling_rms
-from resurfemg.helper_functions import smooth_for_baseline
-from resurfemg.helper_functions import smooth_for_baseline_with_overlay
 
 sample_emg = os.path.join(
     os.path.abspath(os.path.dirname(os.path.dirname(__file__))),
@@ -50,7 +48,6 @@ class TestFilteringMethods(unittest.TestCase):
             (len(sample_emg_filtered[0])),
             len(sample_read.samples[0]) ,
         )
-    
     def test_emg_lowpass_butter(self):
         sample_read= Poly5Reader(sample_emg)
         sample_emg_filtered = emg_lowpass_butter(sample_read.samples, 5, 2048)
@@ -73,7 +70,6 @@ class TestFilteringMethods(unittest.TestCase):
             (len(sample_emg_filtered)),
             len(sample_read.samples[0]) ,
         )
-    
     def test_vect_naive_rolling_rms(self):
         sample_read= Poly5Reader(sample_emg)
         sample_emg_filtered = vect_naive_rolling_rms(sample_read.samples[0], 10)
@@ -114,7 +110,7 @@ class TestComponentPickingMethods(unittest.TestCase):
[…]
