### 2023-10-10 b1a0926
minor text details fixed
