### 2023-10-10 d13f1b0
More linting issues solved
