### 2023-10-10 962456f
More linting issues solved
