### 2023-10-10 bd661af
Second attempt on Config test: changed name and added Rob and mine modifications
