### 2023-10-10 98df0b3
Run all tests in ./tests
