+++ setup.py
@@ -171,16 +171,9 @@ class UnitTest(TestCommand):
             result = runner.run(suite)
             sys.exit(1 if result.errors else 0)
 
-        test_path = os.path.join(project_dir, 'tests')
-        test_files = [test_file.split("\\")[-1] for test_file in glob(
-            test_path + '/*test.py', recursive=False)]
-
-        def execute_tests(test_files):
-            for _, test_file in enumerate(test_files):
-                tests = os.path.join(project_dir, 'tests', test_file)
-                subprocess.call((env_python, '-m', 'unittest', tests))
-
-        sys.exit(execute_tests(test_files))
+        tests = os.path.join(project_dir, 'tests')
+        sys.exit(subprocess.call(
+            (env_python, '-m', 'unittest', 'discover', tests, '*test.py')))
 
 
 class Pep8(TestCommand):
