### 2023-10-10 78cd6b1
Run all tests in ./tests
