### 2023-10-10 58e2611
Restructure repo: config.py to config/config.py + tests
