### 2023-10-10 d7614a7
[fix] Conda no longer tolerates single equals in package version format
