+++ CHANGELOG.md
@@ -116,3 +116,10 @@
 
 	- created due to technical problems with pypi
 	
+## Release 0.1.3
+
+* Version 0.1.3
+
+	- changes in setup.py to accomodate conda package version format
+	- changes to getting started notebook and instructions
+	- required release now (11 October 2023) for JOSS paper 
