### 2023-10-11 8dc3275
Restructure helper_functions into visualisation functions
