### 2023-10-11 bc92e96
Restructure helper_functions into visualisation functions
