### 2023-10-11 #5251 @drcandacemakedamoore
> Fantastic - thanks @drcandacemakedamoore, Can you post a link to the repo and the DOI here? and could check the boxes in the author list [here ](https://github.com/openjournals/joss-reviews/issues/5251#issuecomment-1750742811) to confirm that you've done each step (sorry last request is a bit pedantic!)
>
> Thanks again

I literally can't make more than one check-box check off in every browser I tried. I did all the check-box things though. The repo is https://github.com/resurfemg/resurfemg and the Zenodo DOI is [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8429265.svg)](https://doi.org/10.5281/zenodo.8429265)
