+++ resurfemg/postprocessing/features.py
@@ -1,6 +1,15 @@
+import collections
+import math
+import warnings
+import scipy
 from scipy.fft import fft, fftfreq
+from scipy.signal import savgol_filter
+from scipy.stats import entropy
 import matplotlib.pyplot as plt
 import numpy as np
+import logging
+from ..preprocessing.envelope import running_smoother
+from ..helper_functions.helper_functions import delay_embedding
 
 def show_my_power_spectrum(sample, sample_rate, upper_window):
     """This function plots a power spectrum of the frequencies
@@ -29,3 +38,585 @@ def show_my_power_spectrum(sample, sample_rate, upper_window):
     plt.plot(xf[idx], yf[idx])
     plt.show()
     return yf, xf
+
+
+def entropical(sig):
+    """This function computes something close to certain type of entropy
+    of a series signal array.  Input is sig, the signal, and output is an
+    array of entropy measurements. The function can be used inside a generator
+    to read over slices. Note it is not a true entropy, and works best with
+    very small numbers.
+
+    :param sig: array containin the signal
+    :type sig: ~numpy.ndarray
+
+    :returns: number for an entropy-like signal using math.log w/base 2
+    :rtype: float
+
+    """
+    probabilit = [n_x/len(sig) for x, n_x in collections.Counter(sig).items()]
+    e_x = [-p_x*math.log(p_x, 2) for p_x in probabilit]
+    return sum(e_x)
+
+
+def entropy_scipy(sli, base=None):
+    """
+    This function wraps scipy.stats entropy  (which is a Shannon entropy)
+    for use in the resurfemg library, it can be used in a slice iterator
+    as a drop-in substitute for the hf.entropical but it is a true entropy.
+
+    :param sli: array
+    :type sli: ~numpy.ndarray
+
+    :returns: entropy_count
+    :rtype: float
+    """
+
+    value, counts = np.unique(sli, return_counts=True)
+    entropy_count = entropy(counts/len(counts), base=base)
+    return entropy_count
+
+
+def simple_area_under_curve(
+    array,
+    start_index,
+    end_index,
+):
+    """
+    This function is just a wrapper over np.sum written because it isn't
+    apparent to some clinically oriented people that an area under the curve
+    will be a sum of all the numbers
+
+    :param array: an array e.g. single lead EMG recording
+    :type array: np.array
+    :param start_index: which index number the breath starts on
+    :type start_index: int
+    :param end_index: which index number the breath ends on
+    :type end_index: int
+
+    :returns: area; area under the curve
+    :rtype: float
+    """
+    breath = array[start_index:end_index]
+    area = np.sum(abs(breath))
+    return area
+
+
+def times_under_curve(
+    array,
+    start_index,
+    end_index,
+):
+    """
+    This function is meant to calculate the length of time to peak in
+    an absolute and relative sense
+
+    :param array: an array e.g. single lead EMG recording
+    :type array: np.array
+    :param start_index: which index number the breath starts on
+    :type start_index: int
+    :param end_index: which index number the breath ends on
+    :type end_index: int
+
+    :returns: times; a tuple of absolute and relative times
+    :rtype: tuple
+    """
+    breath_arc = array[start_index:end_index]
+    smoothed_breath = running_smoother(breath_arc)
+    abs_time = smoothed_breath.argmax()
+    percent_time = abs_time / len(breath_arc)
+    times = ((abs_time, percent_time))
+    return times
+
+
+def pseudo_slope(
+    array,
+    start_index,
+    end_index,
+    smoothing=True,
+):
+    """
+    This is a function to get the shape/slope of the take-off angle
+    of the resp. surface EMG signal, however we are returning values of
[…]
