+++ tests/features_test.py
@@ -1,0 +1,114 @@
+#sanity tests for the resurfemg library
+
+
+import unittest
+import numpy as np
+# helper_functions
+from resurfemg.helper_functions import entropy_scipy
+from resurfemg.helper_functions import pseudo_slope
+from resurfemg.helper_functions import area_under_curve
+from resurfemg.helper_functions import simple_area_under_curve
+from resurfemg.helper_functions import times_under_curve
+from resurfemg.helper_functions import find_peak_in_breath
+from resurfemg.helper_functions import variability_maker
+
+
+class TestEntropyMethods(unittest.TestCase):
+
+    def test_entropy_scipy(self):
+        sample_array_lo_entropy = [0,0,0,0,0,0,0,0,0,0]
+        sample_array_hi_entropy = [0,4,0,5,8,0,12,0,1,0]
+        ent_sample_array_lo_entropy = entropy_scipy(sample_array_lo_entropy)
+        ent_sample_array_hi_entropy = entropy_scipy(sample_array_hi_entropy)
+        self.assertGreater(
+            ent_sample_array_hi_entropy ,
+            ent_sample_array_lo_entropy ,
+        )
+
+class TestVariabilityMethods(unittest.TestCase):
+
+    def test_variability_maker_variance(self):
+        sample_array_lo_var = [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0]
+        sample_array_hi_var = [0,4,0,5,8,0,12,0,1,0,0,9,0,9,0,0,9,6,0]
+        var_sample_array_lo_var = variability_maker(sample_array_lo_var, 10)
+        var_sample_array_hi_var = variability_maker(sample_array_hi_var, 10)
+        self.assertGreater(
+            np.sum(var_sample_array_hi_var) ,
+            np.sum(var_sample_array_lo_var) ,
+        )
+
+    def test_variability_maker_std(self):
+        sample_array_lo_var = [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0]
+        sample_array_hi_var = [0,4,0,5,8,0,12,0,1,0,0,9,0,9,0,0,9,6,0]
+        var_sample_array_lo_var = variability_maker(sample_array_lo_var, 10, method='std')
+        var_sample_array_hi_var = variability_maker(sample_array_hi_var, 10, method='std')
+        self.assertGreater(
+            np.sum(var_sample_array_hi_var) ,
+            np.sum(var_sample_array_lo_var) ,
+        )
+
+class TestArrayMath(unittest.TestCase):
+    
+
+    def test_simple_area_under_curve(self):
+        sample_array= np.array(
+            [0,0,0,0,1,1,1,-5,10,10,0,0,0,1,1,1,0,1,1,1,0,0,0]
+        )
+        counted = simple_area_under_curve(sample_array,0,10)
+        self.assertEqual(
+            counted,
+            28,
+        )
+
+    def test_area_under_curve(self):
+        sample_array= np.array(
+            [0,0,0,0,1,1,1,5,10,10,5,0,1,1,1,1,0,1,1,1,0,0,0]
+        )
+        counted = area_under_curve(sample_array,0,20,70)
+        self.assertEqual(
+            counted,
+            28,
+        )
+    
+    def test_times_under_curve(self):
+        sample_array= np.array(
+            [0,1,2,3,1,5,6,-5,8,9,20,11,12,13,4,5,6,1,1,1,0]
+        )
+        counted = times_under_curve(sample_array,0,20)
+        self.assertEqual(
+            counted,
+            ((10,0.5)),
+        )
+
+    def test_pseudo_slope(self):
[…]
