+++ tests/test.py
@@ -7,7 +7,7 @@ import numpy as np
 import scipy
 from tempfile import TemporaryDirectory
 import json
-from unittest import TestCase, main
+# from unittest import TestCase, main
 
 # tmsisdk_lite
 from resurfemg.data_connector.tmsisdk_lite import Poly5Reader
@@ -37,24 +37,24 @@ from resurfemg.helper_functions import zero_one_for_jumps_base
 # from resurfemg.helper_functions import compute_ICA_two_comp
 # from resurfemg.helper_functions import compute_ICA_two_comp_multi
 from resurfemg.helper_functions import working_pipeline_exp
-from resurfemg.helper_functions import entropical
-from resurfemg.helper_functions import entropy_scipy
+# from resurfemg.helper_functions import entropical
+# from resurfemg.helper_functions import entropy_scipy
 # from resurfemg.helper_functions import smooth_for_baseline
 # from resurfemg.helper_functions import smooth_for_baseline_with_overlay
 from resurfemg.helper_functions import relative_levenshtein
 # from resurfemg.helper_functions import gating
 from resurfemg.helper_functions import scale_arrays
-from resurfemg.helper_functions import pseudo_slope
-from resurfemg.helper_functions import area_under_curve
-from resurfemg.helper_functions import simple_area_under_curve
-from resurfemg.helper_functions import times_under_curve
-from resurfemg.helper_functions import find_peak_in_breath
+# from resurfemg.helper_functions import pseudo_slope
+# from resurfemg.helper_functions import area_under_curve
+# from resurfemg.helper_functions import simple_area_under_curve
+# from resurfemg.helper_functions import times_under_curve
+# from resurfemg.helper_functions import find_peak_in_breath
 from resurfemg.helper_functions import distance_matrix
 # from resurfemg.helper_functions import emg_lowpass_butter
 from resurfemg.preprocessing.ecg_removal import find_peaks_in_ecg_signal
-from resurfemg.helper_functions import variability_maker
+# from resurfemg.helper_functions import variability_maker
 # config
-from resurfemg.config.config import Config
+# from resurfemg.config.config import Config
 from resurfemg.config.config import make_realistic_syn_emg
 # ml
 from resurfemg.ml import save_ml_output
@@ -145,39 +145,39 @@ class TestHashMethods(unittest.TestCase):
     #     )
 
 
-class TestEntropyMethods(unittest.TestCase):
+# class TestEntropyMethods(unittest.TestCase):
     
-    def test_entropy_scipy(self):
-        sample_array_lo_entropy = [0,0,0,0,0,0,0,0,0,0]
-        sample_array_hi_entropy = [0,4,0,5,8,0,12,0,1,0]
-        ent_sample_array_lo_entropy = entropy_scipy(sample_array_lo_entropy)
-        ent_sample_array_hi_entropy = entropy_scipy(sample_array_hi_entropy)
-        self.assertGreater(
-            ent_sample_array_hi_entropy ,
-            ent_sample_array_lo_entropy ,
-        )
+#     def test_entropy_scipy(self):
+#         sample_array_lo_entropy = [0,0,0,0,0,0,0,0,0,0]
+#         sample_array_hi_entropy = [0,4,0,5,8,0,12,0,1,0]
+#         ent_sample_array_lo_entropy = entropy_scipy(sample_array_lo_entropy)
[…]
