### 2023-10-13 070400f
helper functions new folder and script
