+++ tests/test.py
@@ -348,13 +348,13 @@ class TestPipelineMethods(unittest.TestCase):
 
 class TestVentCompareMethods(unittest.TestCase):
 
-    def test_relative_levenshtein(self):
-        array1 = np.array([1,0,1,0,1,0])
-        array2 = np.array([1,0,1,0,1,0])
-        array3 = np.array([1,0,1,0,1,0])
-        our_result12 = (relative_levenshtein(array1,array2))
-        our_result13 = (relative_levenshtein(array1,array3))
-        self.assertEqual(our_result12, our_result13)
+    # def test_relative_levenshtein(self):
+    #     array1 = np.array([1,0,1,0,1,0])
+    #     array2 = np.array([1,0,1,0,1,0])
+    #     array3 = np.array([1,0,1,0,1,0])
+    #     our_result12 = (relative_levenshtein(array1,array2))
+    #     our_result13 = (relative_levenshtein(array1,array3))
+    #     self.assertEqual(our_result12, our_result13)
 
 # class TestGating(unittest.TestCase):
 #     sample_read= Poly5Reader(sample_emg)
@@ -425,24 +425,24 @@ class TestMl(unittest.TestCase):
 class TestArrayMath(unittest.TestCase):
     
     
-    def test_scale_arrays(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 500)
-        new_emg = scale_arrays(sample_emg_filtered , 3,0)
-        self.assertEqual(
-            (new_emg.shape),
-            (sample_emg_filtered.shape),
-        )
+    # def test_scale_arrays(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 500)
+    #     new_emg = scale_arrays(sample_emg_filtered , 3,0)
+    #     self.assertEqual(
+    #         (new_emg.shape),
+    #         (sample_emg_filtered.shape),
+    #     )
 
-    def test_zero_one_for_jumps_base(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 500)
-        new_emg = zero_one_for_jumps_base(sample_emg_filtered[0] , sample_emg_filtered[0].mean())
-        new_emg = np.array(np.vstack((new_emg, new_emg)))
-        self.assertEqual(
-            (new_emg.shape[1]),
-            (sample_emg_filtered.shape[1]),
-        )
+    # def test_zero_one_for_jumps_base(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 500)
+    #     new_emg = zero_one_for_jumps_base(sample_emg_filtered[0] , sample_emg_filtered[0].mean())
+    #     new_emg = np.array(np.vstack((new_emg, new_emg)))
+    #     self.assertEqual(
+    #         (new_emg.shape[1]),
+    #         (sample_emg_filtered.shape[1]),
+    #     )
 
     def test_find_peaks_in_ecg_signal(self):
         samp_array = np.array([0,0,0,0,10,0,0,0,10,0,0,0,4,0,0,])
@@ -452,13 +452,13 @@ class TestArrayMath(unittest.TestCase):
             2,
         )
 
-    def test_count_decision_array(self):
-        sample_array= np.array([0,0,0,0,1,1,1,0,0,0,0,0,1,1,1,1,0,1,1,1,0,0,0])
-        counted = count_decision_array(sample_array)
-        self.assertEqual(
-            counted,
-             3,
-        )
[…]
