### 2023-10-17 a9a627a
fix  check and rename helper_functions
