+++ resurfemg/helper_functions_legacy.py
@@ -9,7 +9,7 @@ and other types of data arrays e.g. ventilator signals.
 
 
 import collections
-from collections import namedtuple
+#from collections import namedtuple
 import math
 import warnings
 import time
@@ -25,8 +25,8 @@ from scipy.stats import entropy
 # import matplotlib.pyplot as plt
 import numpy as np
 from sklearn.decomposition import FastICA
-import textdistance
-import pandas as pd
+#import textdistance
+#import pandas as pd
 import logging
 from .preprocessing.filtering import bad_end_cutter
 from .preprocessing.filtering import bad_end_cutter_for_samples
@@ -36,52 +36,52 @@ from .preprocessing.ecg_removal import pick_more_peaks_array
 from .preprocessing.ecg_removal import pick_lowest_correlation_array
 
 
-class Range(namedtuple('RangeBase', 'start,end')):
+# class Range(namedtuple('RangeBase', 'start,end')):
 
-    """Utility class for working with ranges (intervals).
+#     """Utility class for working with ranges (intervals).
 
-    :ivar start: Start of the range
-    :type start: ~number.Number
-    :ivar end: End of the range
-    :type end: ~number.Number
+#     :ivar start: Start of the range
+#     :type start: ~number.Number
+#     :ivar end: End of the range
+#     :type end: ~number.Number
 
-    """
+#     """
 
-    def intersects(self, other):
-        """Returns :code:`True` if this range intersects :code:`other` range.
+#     def intersects(self, other):
+#         """Returns :code:`True` if this range intersects :code:`other` range.
 
-        :param other: Another range to compare this one to
-        :type other: ~resurfemg.helper_functions.Range
+#         :param other: Another range to compare this one to
+#         :type other: ~resurfemg.helper_functions.Range
 
-        :returns: :code:`True` if this range intersects another range
-        :rtype: bool
-        """
-        return (
-            (self.end >= other.end) and (self.start < other.end) or
-            (self.end >= other.start) and (self.start < other.start) or
-            (self.end < other.end) and (self.start >= other.start)
-        )
+#         :returns: :code:`True` if this range intersects another range
+#         :rtype: bool
+#         """
+#         return (
+#             (self.end >= other.end) and (self.start < other.end) or
+#             (self.end >= other.start) and (self.start < other.start) or
+#             (self.end < other.end) and (self.start >= other.start)
+#         )
 
-    def precedes(self, other):
-        """Returns :code:`True` if this range precedes :code:`other` range.
+#     def precedes(self, other):
+#         """Returns :code:`True` if this range precedes :code:`other` range.
 
-        :param other: Another range to compare this one to
-        :type other: ~resurfemg.helper_functions.Range
+#         :param other: Another range to compare this one to
+#         :type other: ~resurfemg.helper_functions.Range
 
-        :returns: :code:`True` if this range strictly precedes another
-            range
-        :rtype: bool
-        """
-        return self.end < other.start
+#         :returns: :code:`True` if this range strictly precedes another
+#             range
+#         :rtype: bool
+#         """
+#         return self.end < other.start
 
-    def to_slice(self):
-        """Converts this range to a :class:`slice`.
+#     def to_slice(self):
+#         """Converts this range to a :class:`slice`.
 
-        :returns: A slice with its start set to this range's start and
-            end set to this range's end
-        :rtype: slice
-        """
-        return slice(*map(int, self))   # maps whole tuple set
[…]
