### 2023-10-17 2938779
change helper functions split and rename legacy
