### 2023-10-17 ccdfc0c
change helper functions split and rename legacy
