+++ paper/paper.bib
@@ -101,7 +101,8 @@ doi={10.1007/978-3-642-34546-3_71}
   number={2},
   pages={118--134},
   year={2013},
-  publisher={C{\`a}tedra d'Accessibilitat (CATAC)}
+  publisher={C{\`a}tedra d'Accessibilitat (CATAC)},
+  url = "https://www.raco.cat/index.php/JACCES/article/view/315915"
 }
 
 @Article{Makowski2021,
