### 2023-10-19 e84768f
paper changes 2 cities
