+++ tests/test.py
@@ -12,12 +12,12 @@ import json
 # tmsisdk_lite
 from resurfemg.data_connector.tmsisdk_lite import Poly5Reader
 # converter_functions 
-from resurfemg.data_connector.converter_functions import poly5unpad
-from resurfemg.config.config import hash_it_up_right_all
+# from resurfemg.data_connector.converter_functions import poly5unpad
+# from resurfemg.config.config import hash_it_up_right_all
 # multi_lead_type
-from resurfemg.multi_lead_type import compute_ICA_n_comp
-from resurfemg.multi_lead_type import compute_ICA_n_comp_selective_zeroing
-from resurfemg.multi_lead_type import compute_ICA_two_comp_selective
+# from resurfemg.multi_lead_type import compute_ICA_n_comp
+# from resurfemg.multi_lead_type import compute_ICA_n_comp_selective_zeroing
+# from resurfemg.multi_lead_type import compute_ICA_two_comp_selective
 # from resurfemg.multi_lead_type import pick_highest_correlation_array_multi
 from resurfemg.multi_lead_type import working_pipe_multi
 from resurfemg.multi_lead_type import working_pipeline_pre_ml_multi
@@ -25,33 +25,33 @@ from resurfemg.multi_lead_type import working_pipeline_pre_ml_multi
 # from resurfemg.helper_functions import bad_end_cutter
 # from resurfemg.helper_functions import bad_end_cutter_better
 # from resurfemg.helper_functions import bad_end_cutter_for_samples
-from resurfemg.helper_functions.helper_functions import count_decision_array
+# from resurfemg.helper_functions.helper_functions import count_decision_array
 from resurfemg.preprocessing.filtering  import emg_bandpass_butter
-from resurfemg.preprocessing.filtering import emg_bandpass_butter_sample
+# from resurfemg.preprocessing.filtering import emg_bandpass_butter_sample
 # from resurfemg.helper_functions import notch_filter
 # from resurfemg.helper_functions import naive_rolling_rms
 # from resurfemg.helper_functions import vect_naive_rolling_rms
-from resurfemg.preprocessing.ecg_removal import pick_more_peaks_array
+# from resurfemg.preprocessing.ecg_removal import pick_more_peaks_array
 # from resurfemg.helper_functions import pick_lowest_correlation_array
-from resurfemg.helper_functions.helper_functions import zero_one_for_jumps_base
+# from resurfemg.helper_functions.helper_functions import zero_one_for_jumps_base
 # from resurfemg.helper_functions import compute_ICA_two_comp
 # from resurfemg.helper_functions import compute_ICA_two_comp_multi
-from resurfemg.helper_functions_legacy import working_pipeline_exp
+# from resurfemg.helper_functions_legacy import working_pipeline_exp
 # from resurfemg.helper_functions import entropical
 # from resurfemg.helper_functions import entropy_scipy
 # from resurfemg.helper_functions import smooth_for_baseline
 # from resurfemg.helper_functions import smooth_for_baseline_with_overlay
-from resurfemg.helper_functions.helper_functions import relative_levenshtein
+# from resurfemg.helper_functions.helper_functions import relative_levenshtein
 # from resurfemg.helper_functions import gating
-from resurfemg.helper_functions.helper_functions import scale_arrays
+# from resurfemg.helper_functions.helper_functions import scale_arrays
 # from resurfemg.helper_functions import pseudo_slope
[…]
